Move command files to new folder
