+++ ethpm_cli/commands/install.py
@@ -12,10 +12,10 @@ from ethpm.backends.ipfs import BaseIPFSBackend
 from ethpm.uri import is_ipfs_uri
 
 from ethpm_cli._utils.filesystem import atomic_replace, is_package_installed
+from ethpm_cli.commands.package import Package
 from ethpm_cli.config import Config
 from ethpm_cli.constants import ETHPM_PACKAGES_DIR, LOCKFILE_NAME, SRC_DIR_NAME
 from ethpm_cli.exceptions import InstallError
-from ethpm_cli.package import Package
 from ethpm_cli.validation import validate_parent_directory
 
 logger = logging.getLogger("ethpm_cli.install")
