+++ tests/core/test_install.py
@@ -4,14 +4,14 @@ import logging
 import pytest
 
 from ethpm_cli._utils.filesystem import check_dir_trees_equal
-from ethpm_cli.constants import ETHPM_PACKAGES_DIR
-from ethpm_cli.exceptions import InstallError
-from ethpm_cli.install import (
+from ethpm_cli.commands.install import (
     install_package,
     list_installed_packages,
     uninstall_package,
 )
-from ethpm_cli.package import Package
+from ethpm_cli.commands.package import Package
+from ethpm_cli.constants import ETHPM_PACKAGES_DIR
+from ethpm_cli.exceptions import InstallError
 
 
 @pytest.fixture
